moved all to cpagent
